Thêm bước xác thực khi xoá học viên

+++ xoa_hoc_vien.py
@@ -4,7 +4,7 @@ import file_handling
 
 
 def xoa_hoc_vien():
-    print("[ - Xoá học viên. - ]")
+    print("[ - Xoá học viên - ]")
     file_hoc_vien = open(file_handling.get_file_path(), "r")
     danh_sach = file_hoc_vien.readlines()
     file_path = file_hoc_vien.name
@@ -19,18 +19,22 @@ def xoa_hoc_vien():
             mshv_entry = entry.split(sep="|")
 
             if mshv_entry[0] == mshv_input:
-                print(f"Đã tìm thấy {mshv_input}.")
+                print(f"\nĐã tìm thấy {mshv_input}.")
                 danh_sach.remove(entry)
                 found = True
                 file_hoc_vien.close()
             else:
                 continue
-
         if found:
-            file_modded = open(file_path, "w")
-            for entries in danh_sach:
-                file_modded.write(entries)
-            file_modded.close()
-            print("Đã xoá", mshv_input, "khỏi danh sách.")
+            print(f"\nBạn có chắc chắn muốn xoá học viên {mshv_input} không?")
+            delete_confirmation = input("Gõ Y để xoá, N để huỷ: ")
+            if delete_confirmation == "Y":
+                file_modded = open(file_path, "w")
+                for entries in danh_sach:
+                    file_modded.write(entries)
+                file_modded.close()
+                print("\nĐã xoá", mshv_input, "khỏi danh sách.")
+            elif delete_confirmation == "N":
+                print("\nĐã huỷ xoá.")
         elif not found:
-            print("Không thể tìm thấy mã số học viên", mshv_input, "trong danh sách.")
+            print("\nKhông thể tìm thấy mã số học viên", mshv_input, "trong danh sách.")
